% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_aggregation.R
\name{aggregate_count}
\alias{aggregate_count}
\title{Aggregating by counts}
\usage{
aggregate_count(df, split.by = NULL, ignore.missing.data = T,
  write.to.file = NULL)
}
\arguments{
\item{df}{data frame to aggregate}

\item{split.by}{aggregating in groups. Creates one row per unique value in the data column with the name speficied here}

\item{ignore.missing.data}{if TRUE, NA and empty strings are not counted as a category}

\item{write.to.file}{specify a CSV file name to write results to}
}
\description{
Aggregating by counts
}
\seealso{
\code{\link{function_name}}
}
