% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_aggregation.R
\name{aggregate_count_weighted}
\alias{aggregate_count_weighted}
\title{Aggregating by weighted counts}
\usage{
aggregate_count_weighted(df, weight.by = NULL, split.by = NULL,
  ignore.missing.data = T, write.to.file = NULL)
}
\description{
Aggregating by weighted counts
}
\examples{



}
\seealso{
\code{\link{function_name}}
}
