% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_aggregation.R
\name{aggregate_frequent}
\alias{aggregate_frequent}
\title{Aggregating to most frequent values}
\usage{
aggregate_frequent(df, n = 3, split.by = NULL, write.to.file = NULL)
}
\description{
Aggregating to most frequent values
}
\examples{

}
\seealso{
\code{\link{function_name}}
}
