% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_aggregation.R
\name{aggregate_frequent_weighted}
\alias{aggregate_frequent_weighted}
\title{Aggregate most frequent values with stratification weights}
\usage{
aggregate_frequent_weighted(df, n = 3, split.by = NULL, weight.by,
  write.to.file = NULL)
}
\description{
Aggregate most frequent values with stratification weights
}
\examples{



}
\seealso{
\code{\link{function_name}}
}
