% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_aggregation.R
\name{aggregate_median}
\alias{aggregate_median}
\title{Aggregating by median}
\usage{
aggregate_median(df, aggregate_by, write.to.file = NULL)
}
\description{
Aggregating by median
}
\examples{



}
\seealso{
\code{\link{function_name}}
}
