% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_aggregation.R
\name{aggregate_percent}
\alias{aggregate_percent}
\title{Aggregating to percentages}
\usage{
aggregate_percent(df, split.by = NULL, ignore.missing.data = TRUE,
  write.to.file = NULL)
}
\description{
Aggregating to percentages
}
\examples{



}
\seealso{
\code{\link{function_name}}
}
