% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_aggregation.R
\name{aggregate_percent_weighted}
\alias{aggregate_percent_weighted}
\title{Aggregating to weighted percentages}
\usage{
aggregate_percent_weighted(df, weight.by = NULL, split.by = NULL,
  ignore.missing.data = T, write.to.file = NULL)
}
\description{
Aggregating to weighted percentages
}
\examples{



}
\seealso{
\code{\link{function_name}}
}
