% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cleaning_check.R
\name{data_validation_all_others}
\alias{data_validation_all_others}
\title{frequencies of all answers in "other" columns}
\usage{
data_validation_all_others(data)
}
\arguments{
\item{data}{data}
}
\value{
data frame with one row per unique "other" response per variable.
}
\description{
In standard format of all data_validation...() functions
}
