% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_load.R
\name{data_validation_cleaning_log_comparison}
\alias{data_validation_cleaning_log_comparison}
\title{Compare cleaning log and cleaned data}
\usage{
data_validation_cleaning_log_comparison(data, data.uuid.column)
}
\arguments{
\item{data}{data frame with the cleaned data}

\item{data.uuid.column}{name of the data column holding the UUIDs}
}
\value{
- Data frame containing uuids, variables and values of d
ata entries that do not match the cleaning log
`NULL` if `return.stratum.populations` is `FALSE` (default)
A named vector with the population counts per stratum, if `return.stratum.populations` is set to `TRUE`
}
\description{
Compare cleaning log and cleaned data
}
\seealso{
\code{\link{load_cleaninglog}} must be run before this function can be used.
}
