% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cleaning_check.R
\name{data_validation_find_duplicates}
\alias{data_validation_find_duplicates}
\title{Lists duplicates in a dataframe column
Returns all numerical values more than three standard deviations from the mean}
\usage{
data_validation_find_duplicates(data, duplicate.column.name)
}
\arguments{
\item{data}{dataframe}

\item{duplicate.column.name}{name of the column to be searched for duplicate entries}
}
\value{
returns values and counts (as a matrix) of duplicate entries in duplicate.column.name
}
\description{
Lists duplicates in a dataframe column
Returns all numerical values more than three standard deviations from the mean
}
