% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cleaning_check.R
\name{data_validation_outliers}
\alias{data_validation_outliers}
\title{Lists outliers in a dataframe
Returns all numerical values more than three standard deviations from the mean (assumes normal distribution)}
\usage{
data_validation_outliers(data, maximum_standard_deviations = 3)
}
\arguments{
\item{data}{dataframe}

\item{maximum_standard_deviations}{set the distance from the mean threshold for outliers in number of standard deviations. defaults to 3}
}
\value{
data frame with the outlier values, as well as the row index and the name of the data column in which they appeared
}
\description{
Lists outliers in a dataframe
Returns all numerical values more than three standard deviations from the mean (assumes normal distribution)
}
