% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_hidden.R
\name{hasdata}
\alias{hasdata}
\title{has data
removes NA, empty strings and non-finite values from a vector}
\usage{
hasdata(x, return.index = F)
}
\arguments{
\item{x}{vector}

\item{return.index}{if true, returns indices of the vector that have valid data. Defaults to FALSE.}
}
\value{
returns the values of the input vector that contain valid data
}
\description{
has data
removes NA, empty strings and non-finite values from a vector
}
\examples{
example1 code
example1 code
}
\seealso{
\code{\link{}}
}
