% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_load.R
\name{load_cleaninglog}
\alias{load_cleaninglog}
\title{Loads the cleaning log from a csv file}
\usage{
load_cleaninglog(cleaning_log_file, cleaning.log.uuid.column,
  cleaning.log.new.value.column, cleaning.log.variable.column)
}
\arguments{
\item{cleaning_log_file}{name of a csv file containing the cleaning log}

\item{cleaning.log.uuid.column}{name of the cleaning log column holding the UUIDs}

\item{cleaning.log.new.value.column}{name of the cleaning log column holding the new value}

\item{cleaning.log.variable.column}{name of the cleaning log column holding the variable name, referencing the column header in the data}
}
\value{
- makes function \code{\link{data_validation_cleaning_log_comparison}} usable
`NULL` if `return.stratum.populations` is `FALSE` (default)
A named vector with the population counts per stratum, if `return.stratum.populations` is set to `TRUE`
}
\description{
Loads the cleaning log from a csv file
}
\seealso{
This function enables \code{\link{data_validation_cleaning_log_comparison}. Similar data loading functions see \link{load_data},\link{load_questionnaire} (not implemented)}
}
