% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_load.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(file, uuid.column = "_uuid")
}
\arguments{
\item{file}{name of a csv file containing the data}

\item{uuid.column}{name of the data colum containing the UUID's}
}
\value{
Data frame with the data
}
\description{
load_data
}
\examples{
load_data("mydata.csv",uuid.column="UUID")
}
\seealso{
\code{\link{load.questionnaire} (not implemented), \link{load_samplingframe}}
}
