% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_questionnaire.R
\name{load_questionnaire}
\alias{load_questionnaire}
\title{load_questionnaire}
\usage{
load_questionnaire(data, questions.file, choices.file,
  choices.label.column.to.use = "label..English")
}
\arguments{
\item{questions.file}{file name of a csv file containing the kobo form's question sheet}

\item{choices.file}{file name of a csv file containing the kobo form's choices sheet}
}
\value{
list of questions and choices, sorted to match. to data columns.
}
\description{
load_questionnaire
}
\examples{

}
\seealso{
\code{\link{load_data()}, \link{load_samplingframe}}
}
