% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_load.R
\name{load_samplingframe}
\alias{load_samplingframe}
\title{Loads the sampling frame from a csv file, performing inital tests on the data.}
\usage{
load_samplingframe(sampling.frame.file,
  sampling.frame.population.column = "population",
  sampling.frame.stratum.column = "stratum",
  data.stratum.column = "stratum", return.stratum.populations = FALSE)
}
\arguments{
\item{sampling.frame.file}{name of a csv file containing the sampling frame}

\item{sampling.frame.population.column}{sampling frame name of column holding population counts}

\item{sampling.frame.stratum.column}{sampling frame name of column holding stratum names. Stratum names must match exactly values in:}

\item{data.stratum.column}{data column name that holds the record's strata names}

\item{return.stratum.populations}{by default this function returns `NULL`, but must be called to make automatic weighting in ..._weighted() functions possible. you can however retreive the stratum population counts by setting this to `TRUE`.}
}
\value{
- makes function `weights_of()` usable, and with it .._weighted() functions with automatic weighting
`NULL` if `return.stratum.populations` is `FALSE` (default)
A named vector with the population counts per stratum, if `return.stratum.populations` is set to `TRUE`
}
\description{
Loads the sampling frame from a csv file, performing inital tests on the data.
}
\examples{
load_data("mydata.csv",uuid.column="UUID")
}
\seealso{
\code{\link{load_data} (not implemented), \link{aggregate_count_weighted}}
}
