% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_hidden.R
\name{medianORmode}
\alias{medianORmode}
\title{Median or Mode}
\usage{
medianORmode(x, when.tie = NA, mean.when.fraction.numeric.greater.than = 1)
}
\description{
Median or Mode
}
\examples{
mixed_data <- c(1,2,3,"4","5","a","b")
medianORmode(mixed_data,"no median consensus",0.5)
medianORmode(mixed_data,"no median consensus",0.9)
}
\seealso{
\code{\link{function_name}}
}
