% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_aggregation.R
\name{percent_increase}
\alias{percent_increase}
\title{percent_increase}
\usage{
percent_increase(from, to, match.column)
}
\description{
percent_increase
}
\examples{



}
\seealso{
\code{\link{function_name}}
}
