% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_questionnaire.R
\name{question_get_choice_labels}
\alias{question_get_choice_labels}
\title{question_get_choice_labels}
\usage{
question_get_choice_labels(responses, variable.name)
}
\arguments{
\item{responses}{a vector of responses}

\item{variable.name}{the column name of the corresponding variable}
}
\value{
vector of labels corresponding to the responses vector. If no corresponding label is found, return the original value
}
\description{
question_get_choice_labels
}
\examples{

}
