% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_questionnaire.R
\name{question_is_categorical}
\alias{question_is_categorical}
\title{question_is_categorical}
\usage{
question_is_categorical(question.name)
}
\arguments{
\item{question.name}{the name of a data column header}
}
\value{
TRUE if the question is of type select_one or select_multiple
}
\description{
question_is_categorical
}
\examples{

}
