% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_questionnaire.R
\name{question_is_numeric}
\alias{question_is_numeric}
\title{question_is_numeric}
\usage{
question_is_numeric(question.name)
}
\arguments{
\item{question.name}{the name of a data column header}
}
\value{
TRUE if the question is of type integer or decimal
}
\description{
question_is_numeric
}
\examples{

}
