% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_questionnaire.R
\name{question_is_select_multiple}
\alias{question_is_select_multiple}
\title{question_is_select_multiple}
\usage{
question_is_select_multiple(question.name)
}
\arguments{
\item{question.name}{the name of a data column header}
}
\value{
TRUE if the question is of type select_multiple
}
\description{
question_is_select_multiple
}
\examples{

}
