% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_questionnaire.R
\name{question_is_select_one}
\alias{question_is_select_one}
\title{question_is_select_one}
\usage{
question_is_select_one(question.name)
}
\arguments{
\item{question.name}{the name of a data column header}
}
\value{
TRUE if the question is of type select_one
}
\description{
question_is_select_one
}
\examples{

}
