% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{reach_style_color_beige}
\alias{reach_style_color_beige}
\title{reach brand beiges}
\usage{
reach_style_color_beige(lightness = 1, transparent = F)
}
\description{
reach brand beiges
}
\examples{



}
\seealso{
\code{\link{function_name}}
}
