% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{reach_style_color_beiges}
\alias{reach_style_color_beiges}
\title{Reach brand beige triples}
\usage{
reach_style_color_beiges(transparent = F)
}
\description{
Reach brand beige triples
}
\examples{



}
\seealso{
\code{\link{function_name}}
}
