% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{reach_style_color_red}
\alias{reach_style_color_red}
\title{Reach brand reds}
\usage{
reach_style_color_red(lightness = 1, transparent = F)
}
\description{
Reach brand reds
}
\examples{



}
\seealso{
\code{\link{function_name}}
}
