% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{reach_style_line__stacked}
\alias{reach_style_line__stacked}
\title{Stacked line graph}
\usage{
reach_style_line__stacked(df, x.col.name, y.col.names, xlab, ylab,
  labels = NULL, highlight = NULL, write.to.file = NULL)
}
\description{
Stacked line graph
}
\examples{



}
\seealso{
\code{\link{function_name}}
}
