% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{reach_style_line_with.band}
\alias{reach_style_line_with.band}
\title{Line graph with confidence band}
\usage{
reach_style_line_with.band(x, y, y.low, y.high, xlab = "", ylab = "")
}
\description{
Line graph with confidence band
}
\examples{



}
\seealso{
\code{\link{function_name}}
}
