% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{reach_style_points_by_category_with_error_bars}
\alias{reach_style_points_by_category_with_error_bars}
\title{points with horizontal error bars}
\usage{
reach_style_points_by_category_with_error_bars(group, value, value.low,
  value.high, xlab = "", ylab = "")
}
\description{
points with horizontal error bars
}
\examples{


}
\seealso{
\code{\link{function_name}}
}
